% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lib.R
\docType{class}
\name{ArgentinaNamesRetriever.class}
\alias{ArgentinaNamesRetriever.class}
\title{Names distributions}
\format{\code{\link{R6Class}} object.}
\usage{
ArgentinaNamesRetriever.class
}
\description{
Class for analysing time-distribution of names from a names database with fields: name, quantity, year
}
\section{Fields}{

\describe{
\item{\code{file.id}}{polyhedron file id}
}}

\section{Methods}{

\describe{
  \item{\code{process()}}{}
}
}

\keyword{datasets}
