% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamesDistribution.R
\name{generateSampleDistribution}
\alias{generateSampleDistribution}
\title{Generate Sample Distribution
Gets a subset of}
\usage{
generateSampleDistribution(name.year.count, seed, sample.ratio = 0.01,
  min.count = 100)
}
\description{
Generate Sample Distribution
Gets a subset of
}
