% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamesDistribution.R
\docType{class}
\name{NamesDistribution.class}
\alias{NamesDistribution.class}
\title{Names distributions}
\format{\code{\link{R6Class}} object.}
\usage{
NamesDistribution.class
}
\description{
Class for analysing time-distribution of names from a names database with fields: name, quantity, year
}
\section{Methods}{

\describe{
  \item{\code{process()}}{}
}
}

\keyword{datasets}
