% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamesDistributionSimulation.R
\docType{data}
\name{NamesDistributionSimulation.class}
\alias{NamesDistributionSimulation.class}
\title{Names Distribution Simulation abstract class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamesDistributionSimulation.class
}
\description{
Names Distribution Simulation abstract class
}
\keyword{datasets}
