% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamesDistributionSimulation.R
\docType{data}
\name{NamesDistributionSimulationMultipleRuns.class}
\alias{NamesDistributionSimulationMultipleRuns.class}
\title{Names Distribution Simulation Multiple Runs is a container class of multiple simulations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamesDistributionSimulationMultipleRuns.class
}
\description{
Names Distribution Simulation Multiple Runs is a container class of multiple simulations
}
\keyword{datasets}
