% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamesDistributionSimulation.R
\docType{data}
\name{NamesDistributionSimulationRun.class}
\alias{NamesDistributionSimulationRun.class}
\title{Names Distribution Simulation Run is a container class of a unique simulation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamesDistributionSimulationRun.class
}
\description{
Names Distribution Simulation Run is a container class of a unique simulation
}
\keyword{datasets}
