% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{normalizeString}
\alias{normalizeString}
\title{Normalize String
Remove accents, trim and lower input text}
\usage{
normalizeString(text)
}
\description{
Normalize String
Remove accents, trim and lower input text
}
