% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter-retriever.R
\docType{data}
\name{TwitterRetriever.class}
\alias{TwitterRetriever.class}
\title{TwitterRetriever}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TwitterRetriever.class
}
\description{
TwitterRetriever
}
\keyword{datasets}
