% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter-retriever.R
\docType{data}
\name{TwitterUsers.class}
\alias{TwitterUsers.class}
\title{TwitterUsers.class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TwitterUsers.class
}
\description{
TwitterUsers.class
}
\keyword{datasets}
